Unify ExtVTable per vtable design doc
