+++ vortex-array/src/dtype/extension/mod.rs
@@ -2,31 +2,43 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 //! Extension DTypes, and interfaces for working with extension types (dtypes, scalars, and arrays).
+//!
+//! ## File layout convention
+//!
+//! Each vtable-backed concept `Foo` follows this module structure:
+//!
+//! - `vtable.rs`  — `FooVTable` (the non-object-safe trait users implement)
+//! - `plugin.rs`  — `FooPlugin` (registry trait for deserialization + blanket impl)
+//! - `typed.rs`   — `Foo<V>` (typed wrapper) + `FooInner<V>` + `DynFoo` (private)
+//! - `erased.rs`  — `FooRef` (erased ref + Display/Debug/PartialEq/Hash impls)
+//! - `matcher.rs` — `Matcher` trait + blanket impl for `V: FooVTable`
+
+mod vtable;
+pub use vtable::ExtVTable;
+
+mod plugin;
+pub use plugin::ExtDTypePlugin;
+
+pub(super) mod typed;
+pub use typed::ExtDType;
+
+mod erased;
+pub use erased::ExtDTypeRef;
 
 mod matcher;
 pub use matcher::Matcher;
 
-mod vtable;
-pub use vtable::DynExtVTable;
-pub use vtable::ExtVTable;
-
-mod dtype;
-pub use dtype::*;
-
 /// A unique identifier for an extension type
 pub type ExtId = arcref::ArcRef<str>;
 
-/// Private module to seal [`ExtDTypeImpl`].
-///
-/// Note that this is not strictly necessary since [`ExtDTypeAdapter`] and [`ExtDTypeImpl`] are both
-/// private to this module, this is just for hygiene.
+/// Private module to seal [`typed::DynExtDType`].
 mod sealed {
-    use crate::dtype::extension::ExtDTypeAdapter;
     use crate::dtype::extension::ExtVTable;
+    use crate::dtype::extension::typed::ExtDTypeInner;
 
-    /// Marker trait to prevent external implementations of [`ExtDTypeImpl`].
+    /// Marker trait to prevent external implementations of [`super::typed::DynExtDType`].
     pub(crate) trait Sealed {}
 
-    /// This can be the **only** implementor for [`ExtDTypeImpl`].
-    impl<V: ExtVTable> Sealed for ExtDTypeAdapter<V> {}
+    /// This can be the **only** implementor for [`super::typed::DynExtDType`].
+    impl<V: ExtVTable> Sealed for ExtDTypeInner<V> {}
 }
