+++ vortex-array/src/dtype/serde/proto.rs
@@ -150,7 +150,7 @@ impl TryFrom<&DType> for pb::DType {
                 DType::Extension(e) => DtypeType::Extension(Box::new(pb::Extension {
                     id: e.id().as_ref().into(),
                     storage_dtype: Some(Box::new(e.storage_dtype().try_into()?)),
-                    metadata: Some(e.metadata_erased().serialize()?),
+                    metadata: Some(e.serialize_metadata()?),
                 })),
             }),
         })
